# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Zenith API Documentation",
        "description": "Documentation for Zenith API endpoints",
        "contact": {
            "email": "admin@zenith.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/manage/toko": {
            "get": {
                "tags": [
                    "Toko"
                ],
                "summary": "Check current user's store status",
                "description": "Cek status toko user saat ini.\nBerguna untuk frontend mengetahui apakah user sudah punya toko atau belum.",
                "operationId": "65760d848dfbbdd16503bd7e02ff377c",
                "responses": {
                    "200": {
                        "description": "Store status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "exists"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Toko"
                ],
                "summary": "Create a new store",
                "description": "Menyimpan Toko Baru.",
                "operationId": "9ddd22e537f67dbfdedb47f9efd38dd2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "toko_name",
                                    "deskripsi"
                                ],
                                "properties": {
                                    "toko_name": {
                                        "type": "string",
                                        "example": "My Store"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Best store ever"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Store created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden (Not a seller)"
                    },
                    "409": {
                        "description": "Conflict (Store already exists)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/toko/{id}": {
            "get": {
                "tags": [
                    "Toko"
                ],
                "summary": "Get store details",
                "operationId": "6db826731d1d6ca1d858081a84693ad3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Store ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Store details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Store not found"
                    }
                }
            }
        },
        "/api/shops": {
            "get": {
                "tags": [
                    "Toko"
                ],
                "summary": "Search shops",
                "description": "Search for shops (users with role 'penjual').",
                "operationId": "bb4a8bea830592588360a490102bec5a",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of shops",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/manage/toko/{id}/freeze": {
            "post": {
                "tags": [
                    "Toko"
                ],
                "summary": "Freeze a store",
                "description": "Freeze a store (Admin only).",
                "operationId": "edc80bf5ae262493502bcbb4d4a8049d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Store ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "type": "string",
                                        "example": "Violation of terms"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Store frozen",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Toko berhasil dibekukan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/manage/toko/{id}/unfreeze": {
            "post": {
                "tags": [
                    "Toko"
                ],
                "summary": "Unfreeze a store",
                "description": "Unfreeze a store (Admin only).",
                "operationId": "b3f127425bd786df152514a47e4f52f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Store ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Store unfrozen",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Toko berhasil diaktifkan kembali"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/manage/toko/{id}/appeal": {
            "post": {
                "tags": [
                    "Toko"
                ],
                "summary": "Submit appeal for frozen store",
                "operationId": "56c63be0364e141b9faeaa49979c8f38",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Store ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "appeal_reason"
                                ],
                                "properties": {
                                    "appeal_reason": {
                                        "type": "string",
                                        "example": "I will be good"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Appeal submitted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Banding berhasil diajukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}